Move `EventBus` to `server` package.
Client-side event handing is a separate thing. We cannot have the same event bus on server- and client-side.
