+++ examples/src/main/java/org/spine3/examples/aggregate/server/Application.java
@@ -25,12 +25,12 @@ import com.google.common.util.concurrent.MoreExecutors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spine3.client.CommandRequest;
-import org.spine3.eventbus.EventBus;
 import org.spine3.examples.aggregate.Client;
 import org.spine3.examples.aggregate.OrderId;
 import org.spine3.server.BoundedContext;
 import org.spine3.server.CommandBus;
 import org.spine3.server.CommandStore;
+import org.spine3.server.EventBus;
 import org.spine3.server.storage.StorageFactory;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
 import org.spine3.server.stream.EventStore;
