+++ server/src/main/java/org/spine3/internal/EventHandlerMethod.java
@@ -29,8 +29,8 @@ import com.google.protobuf.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spine3.base.EventContext;
-import org.spine3.eventbus.Subscribe;
 import org.spine3.server.MultiHandler;
+import org.spine3.server.Subscribe;
 import org.spine3.server.util.MethodMap;
 import org.spine3.server.util.Methods;
 import org.spine3.type.EventClass;
