+++ server/src/main/java/org/spine3/server/EventBus.java
@@ -17,7 +17,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.spine3.eventbus;
+package org.spine3.server;
 
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.ImmutableList;
