+++ server/src/test/java/org/spine3/eventbus/EventBusShould.java
@@ -23,6 +23,7 @@ package org.spine3.eventbus;
 import com.google.common.util.concurrent.MoreExecutors;
 import org.junit.Before;
 import org.junit.Test;
+import org.spine3.server.EventBus;
 import org.spine3.server.storage.StorageFactory;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
 import org.spine3.server.stream.EventStore;
