+++ server/src/test/java/org/spine3/server/procman/ProcessManagerShould.java
@@ -27,8 +27,8 @@ import org.junit.Test;
 import org.spine3.base.CommandContext;
 import org.spine3.base.EventContext;
 import org.spine3.base.EventRecord;
-import org.spine3.eventbus.Subscribe;
 import org.spine3.server.Assign;
+import org.spine3.server.Subscribe;
 import org.spine3.test.project.ProjectId;
 import org.spine3.test.project.command.AddTask;
 import org.spine3.test.project.command.CreateProject;
